+++ module/paste/src/main/kotlin/sibyl/module/paste/PasteService.kt
@@ -6,5 +6,5 @@ interface PasteService {
      *
      * @return url of paste
      */
-    suspend fun paste(content: String): String
+    suspend fun paste(content: String): String?
 }