+++ module/reminders/src/main/kotlin/sibyl/module/reminders/RemindCommand.kt
@@ -1,11 +1,7 @@
 package sibyl.module.reminders
 
 import com.github.ajalt.clikt.core.subcommands
-import com.github.ajalt.clikt.parameters.arguments.*
-import org.joda.time.Instant
-import org.joda.time.LocalDateTime
 import sibyl.commands.SibylCommand
-import sibyl.db.Reminders
 
 class RemindCommand(remindersModule: RemindersModule) : SibylCommand(
     name = "remind",