+++ util/paste/hastebin/src/main/kotlin/sibyl/haste/HasteService.kt
@@ -3,6 +3,7 @@ import io.ktor.client.*
 import io.ktor.client.request.*
 import io.ktor.content.*
 import io.ktor.http.*
+import kotlinx.coroutines.withTimeoutOrNull
 import kotlinx.serialization.json.Json
 import mu.KotlinLogging
 import sibyl.module.paste.PasteService
@@ -19,20 +20,23 @@ class HasteService(
         }
     }
 
-    override suspend fun paste(content: String): String {
+    override suspend fun paste(content: String): String? {
 
-        val jsonResponse = httpClient.post<String>("$hasteServer/documents") {
-            body = TextContent(
-                text = content,
-                contentType = ContentType.Text.Plain
-            )
+        val url = withTimeoutOrNull(1000) {
+            val jsonResponse = httpClient.post<String>("$hasteServer/documents") {
+                body = TextContent(
+                    text = content,
+                    contentType = ContentType.Text.Plain
+                )
+            }
+
+            val response = jsonSerializer.decodeFromString(HasteServerResponse.serializer(), jsonResponse)
+
+            val url = "$hasteServer/${response.key}"
+
+            logger.debug { "response: $url" }
+            url
         }
-
-        val response = jsonSerializer.decodeFromString(HasteServerResponse.serializer(), jsonResponse)
-
-        val url = "$hasteServer/${response.key}"
-
-        logger.debug { "response: $url" }
         return url
     }
 }